implement configuration in content profile settings UI
