+++ scripts/apps/search/directives/SavedSearchEditOwnSubscription.ts
@@ -1,8 +1,7 @@
 import {ISavedSearch, updateSubscribers, unsubscribeUser, IUserSubscription} from '../SavedSearch';
 import {IDirectiveScope} from 'types/Angular/DirectiveScope';
 import {CronTimeInterval} from 'types/DataStructures/TimeInterval';
-import {IUser} from 'superdesk-api';
-import {IVocabulary} from 'superdesk-interfaces/Vocabulary';
+import {IUser, IVocabulary} from 'superdesk-api';
 
 interface IScope extends IDirectiveScope<void> {
     vocabularies: Array<IVocabulary>;
