+++ scripts/apps/vocabularies/controllers/VocabularyConfigController.ts
@@ -1,5 +1,5 @@
 import {MEDIA_TYPES, MEDIA_TYPE_KEYS, DEFAULT_SCHEMA, VOCABULARY_SELECTION_TYPES} from '../constants';
-import {IVocabulary, IVocabularyTag} from 'superdesk-interfaces/Vocabulary';
+import {IVocabulary, IVocabularyTag} from 'superdesk-api';
 import {IDirectiveScope} from 'types/Angular/DirectiveScope';
 import {remove, reduce} from 'lodash';
 import {gettext} from 'core/utils';
