+++ scripts/apps/workspace/content/directives/ContentProfileSchemaEditor.ts
@@ -1,10 +1,11 @@
 import {includes} from 'lodash';
 import {getLabelForFieldId} from '../../helpers/getLabelForFieldId';
+import {IArticleField} from 'superdesk-api';
 
 interface IScope extends ng.IScope {
     getEditor3FormattingOptions: (fieldName: string) => Array<string>;
     model: any;
-    fields: any;
+    fields: {[key: string]: IArticleField};
     form: any;
     formattingOptions: Array<string>;
     schemaKeysOrdering: any;
@@ -21,6 +22,7 @@ interface IScope extends ng.IScope {
     onOrderUpdate(locals: { key: string; }): any;
     onDrag(locals: { start: number; end: number; key: string; }): any;
     onToggle(locals: { key: string; dest: string; position: string; }): any;
+    showPreviewConfig(field: IArticleField): boolean;
 }
 
 const HAS_PLAINTEXT_FORMATTING_OPTIONS = Object.freeze({
@@ -195,6 +197,10 @@ export function ContentProfileSchemaEditor(vocabularies) {
             vocabularies.getVocabularies().then((vocabulariesCollection) => {
                 scope.label = (id) => getLabelForFieldId(id, vocabulariesCollection);
             });
+
+            // enable preview config for CVs
+            // we display other fields by default already
+            scope.showPreviewConfig = (field) => field != null && field.field_type == null;
         },
     };
 }
