--- scripts/superdesk-interfaces/Vocabulary.ts
@@ -1,26 +1,0 @@
-export interface IVocabularyTag {
-    text: string;
-}
-
-export interface IVocabulary {
-    _id: string;
-    display_name: string;
-    helper_text: string;
-    popup_width: number;
-    type: string;
-    items: Array<{name: string; qcode: string}>;
-    single_value: boolean;
-    schema_field: string;
-    dependent: boolean;
-    service: {};
-    priority: number;
-    unique_field: string;
-    schema: {};
-    field_type: 'text' | 'media' | 'date' | 'embed' | 'related_content' | 'custom';
-    field_options: {};
-    init_version: number;
-    preffered_items: boolean;
-    tags: Array<IVocabularyTag>;
-    date_shortcuts: Array<{value: number; term: string; label: string}>;
-    custom_field_type: string;
-}
